frontend_backend/requirements.txt
setuptools>=65.0.0
wheel
flask==2.0.1
pillow==9.5.0
torch==2.5.0
torchvision==0.20.0
torchaudio==2.5.0
gunicorn==20.1.0
werkzeug==2.0.3
matplotlib==3.4.3
numpy==1.21.6
requests==2.26.0
python-dotenv==0.20.0

